// src/stores/movies/types.ts
/* eslint-disable no-shadow */
export enum MoviesActionTypes {
  MOVIES_FETCH_REQUEST = '@@MOVIES/FETCH/REQUEST',
  MOVIES_FETCH_SUCCESS = '@@MOVIES/FETCH/SUCCESS',
  MOVIES_FETCH_FAILED = '@@MOVIES/FETCH/FAILED',

  MOVIES_FETCH_DETAIL_REQUEST = '@@MOVIES/FETCH/DETAIL/REQUEST',
  MOVIES_FETCH_DETAIL_SUCCESS = '@@MOVIES/FETCH/DETAIL/SUCCESS',
  MOVIES_FETCH_DETAIL_FAILED = '@@MOVIES/FETCH/DETAIL/FAILED',
}

export default MoviesActionTypes;